database/db_test/requirements.txt
psycopg2-binary  # PostgreSQL adapter for Python
python-dotenv    # For environment variable management